iPhoneだと「type 'int' is not a subtype of type 'double'・・・」エラーが発生するので修正

+++ project/bodquest_2023/lib/infrastructure/model/firestore/rug_weight.dart
@@ -10,7 +10,7 @@ class RugWeight {
   factory RugWeight.fromJson(json) {
     return RugWeight(
       userId: json['userId'],
-      value: json['value'],
+      value: json['value'].toDouble(),
     );
   }
 }
